Add image for Nextcloud server used in acceptance tests
The acceptance tests verify the behavior of an actual Nextcloud server.
Due to security concerns, the public Nextcloud server repository is not
set as "trusted" in Drone (otherwise a malicious pull request could be
used to take over the server), so it is not possible to create Docker
containers from the containers started by Drone. Therefore, the
Nextcloud server must be started as a service by Drone itself.

This image provides a base system in which the Nextcloud server can be
installed from the Git repository cloned by Drone. Moreover, it also
provides a helper "daemon" to control the Nextcloud server; the
acceptance tests communicate with this "daemon" to perform certain
actions that require access to the system in which the Nextcloud server
is running on, like resetting it.

Along with the "daemon" a helper script is also provided to set up the
system as needed for the "daemon" itself and the acceptance tests.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>