Add Dockerfile for acceptance tests
Acceptance tests for Nextcloud server can be run on the PHP built-in web
server, so until now the Docker containers for unit or integration tests
were used for the acceptance tests.

However, the PHP built-in web server is single threaded, so it can not
be used to test certain apps, like for example Nextcloud Talk (as it
uses long polling, and thus requires a web server that can handle
several requests in parallel).

This commit adds a new container to be used by acceptance tests; it is
based on the container for unit tests and adds the Apache web server, so
acceptance tests can be run on the PHP built-in web server or on the
Apache web server as needed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>