Build Qt 5.12.5 from source
Since the previously used PPA doesn't offer binaries yet, we compile Qt
from source.

With: OpenSSL 1.1.1d (TLS 1.3)

Note: On Mac we have to compile Qt from source either way because the official
binaries are not linked against OpenSSL. The default is to use Apple Secure Transport which
sadly lacks some features like TLS 1.3 and client certificates.

So we take the gained knowledge and use it for Linux here too ;)

Signed-off-by: Michael Schuster <michael@schuster.ms>