Use precompiled Qt 5.12.5
The precompiled Qt package has been built with the commands in this Dockerfile.

Since it takes a very long time to compile, the build on Docker Hub fails due to a timeout.

This is why we're going to use our own precompiled version here. You may modify the comments
in this Dockerfile to build your own version on a dedicated build machine.

Signed-off-by: Michael Schuster <michael@schuster.ms>