fix(translationtool): Fallback to English plural form
If a project does not use any plurals (only `t` and no `n`),
then the po file might also not contain the plural header.
In this case we still need a plural form string for our Nextcloud JS output.

So to solve this we just fallback to English (it is currently not used
and will also be overwritten with the correct string as soon as a plural form is used)

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>